service/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
psycopg2-binary==2.9.9
python-dateutil==2.4.2
requests==2.31.0
setuptools==68.2.0
six==1.16.0
urllib3==2.1.0
wheel==0.41.2

